move text oriented funcions and classes to text.py
because util.py is too long
