+++ mock/py/mockbuild/util.py
@@ -14,7 +14,6 @@ import errno
 import fcntl
 from glob import glob
 import grp
-import locale
 import logging
 import os
 import os.path
@@ -39,18 +38,14 @@ import time
 import uuid
 
 import distro
-import jinja2
 
 from mockbuild.mounts import BindMountPoint
 
 from . import exception
+from . import text
 from .trace_decorator import getLog, traceLog
 from .uid import getresuid, setresuid
 from pyroute2 import IPRoute
-# pylint: disable=no-name-in-module
-from collections.abc import MutableMapping
-
-encoding = locale.getpreferredencoding()
 
 _libc = ctypes.cdll.LoadLibrary(None)
 _libc.personality.argtypes = [ctypes.c_ulong]
@@ -108,105 +103,6 @@ def cmd_pretty(cmd):
     if isinstance(cmd, list):
         return ' '.join(pipes.quote(arg) for arg in cmd)
     return cmd
-
-
-def compat_expand_string(string, conf_dict):
-    """
-    Expand %(uid)s, etc., only if needed - and warn the user
-    that Jinja should be used instead.
-    """
-    if '%(' not in string:
-        return string
-    getLog().warning("Obsoleted %(foo) config expansion in '{}', "
-                     "use Jinja alternative {{foo}}".format(string))
-    return string % conf_dict
-
-
-# pylint: disable=no-member,unsupported-assignment-operation
-class TemplatedDictionary(MutableMapping):
-    """ Dictionary where __getitem__() is run through Jinja2 template """
-    def __init__(self, *args, alias_spec=None, **kwargs):
-        '''
-        Use the object dict.
-
-        Optional parameter 'alias_spec' is dictionary of form:
-        {'aliased_to': ['alias_one', 'alias_two', ...], ...}
-        When specified, and one of the aliases is accessed - the
-        'aliased_to' config option is returned.
-        '''
-        self.__dict__.update(*args, **kwargs)
-
-        self._aliases = {}
-        if alias_spec:
-            for aliased_to, aliases in alias_spec.items():
-                for alias in aliases:
-                    self._aliases[alias] = aliased_to
-
-    # The next five methods are requirements of the ABC.
-    def __setitem__(self, key, value):
-        key = self._aliases.get(key, key)
-        self.__dict__[key] = value
-    def __getitem__(self, key):
-        key = self._aliases.get(key, key)
-        if '__jinja_expand' in self.__dict__ and self.__dict__['__jinja_expand']:
-            return self.__render_value(self.__dict__[key])
-        return self.__dict__[key]
-    def __delitem__(self, key):
-        del self.__dict__[key]
-    def __iter__(self):
-        return iter(self.__dict__)
-    def __len__(self):
-        return len(self.__dict__)
-    # The final two methods aren't required, but nice to have
-    def __str__(self):
-        '''returns simple dict representation of the mapping'''
-        return str(self.__dict__)
-    def __repr__(self):
-        '''echoes class, id, & reproducible representation in the REPL'''
-        return '{}, TemplatedDictionary({})'.format(super(TemplatedDictionary, self).__repr__(),
-                                                    self.__dict__)
-    def copy(self):
-        return TemplatedDictionary(self.__dict__)
-    def __render_value(self, value):
-        if isinstance(value, str):
-            return self.__render_string(value)
-        elif isinstance(value, list):
-            # we cannot use list comprehension here, as we need to NOT modify the list (pointer to list)
-            # and we need to modifiy only individual values in the list
-            # If we would create new list, we cannot assign to it, which often happens in configs (e.g. plugins)
-            for i in range(len(value)): # pylint: disable=consider-using-enumerate
-                value[i] = self.__render_value(value[i])
-            return value
-        elif isinstance(value, dict):
-            # we cannot use list comprehension here, same reasoning as for `list` above
-            for k in value.keys():
-                value[k] = self.__render_value(value[k])
-            return value
-        else:
-            return value
-    def __render_string(self, value):
-        orig = last = value
-        max_recursion = self.__dict__.get('jinja_max_recursion', 5)
-        for _ in range(max_recursion):
-            template = jinja2.Template(value, keep_trailing_newline=True)
-            value = _to_native(template.render(self.__dict__))
-            if value == last:
-                return value
-            last = value
-        raise ValueError("too deep jinja re-evaluation on '{}'".format(orig))
-
-
-def _to_text(obj, arg_encoding='utf-8', errors='strict', nonstring='strict'):
-    if isinstance(obj, str):
-        return obj
-    elif isinstance(obj, bytes):
-        return obj.decode(arg_encoding, errors)
-    else:
-        if nonstring == 'strict':
-            raise TypeError('First argument must be a string')
-        raise ValueError('nonstring must be one of: ["strict",]')
-
-_to_native = _to_text
 
 
 @traceLog()
@@ -428,7 +324,7 @@ def getNEVRA(hdr):
     if epoch is None:
         epoch = 0
     ret = (name, epoch, ver, rel, arch)
-    return tuple(_to_text(x) if i != 1 else x for i, x in enumerate(ret))
+    return tuple(text._to_text(x) if i != 1 else x for i, x in enumerate(ret))
 
 
 @traceLog()
@@ -614,7 +510,7 @@ def logOutput(fdout, fderr, logger, returnOutput=1, start=0, timeout=0, printOut
                 if returnStderr is False and s == fderr:
                     continue
 
-                txt_input = raw.decode(encoding, 'replace')
+                txt_input = raw.decode(text.encoding, 'replace')
                 lines = txt_input.split("\n")
                 if tail:
                     lines[0] = tail + lines[0]
@@ -1036,7 +932,7 @@ def nspawn_supported():
 @traceLog()
 def setup_default_config_opts(unprivUid, version, pkgpythondir):
     "sets up default configuration."
-    config_opts = TemplatedDictionary(alias_spec={'dnf.conf': ['yum.conf']})
+    config_opts = text.TemplatedDictionary(alias_spec={'dnf.conf': ['yum.conf']})
     config_opts['config_paths'] = []
     config_opts['version'] = version
     config_opts['basedir'] = '/var/lib/mock'  # root name is automatically added to this
@@ -1474,7 +1370,7 @@ def set_config_opts_per_cmdline(config_opts, options, args):
     check_config(config_opts)
     # can't cleanup unless resultdir is separate from the root dir
     basechrootdir = os.path.join(config_opts['basedir'], config_opts['root'])
-    config_resultdir = compat_expand_string(config_opts['resultdir'], config_opts)
+    config_resultdir = text.compat_expand_string(config_opts['resultdir'], config_opts)
     if is_in_dir(config_resultdir, basechrootdir):
         config_opts['cleanup_on_success'] = False
         config_opts['cleanup_on_failure'] = False
